+++ src/dirtygpt/header.nim
@@ -5,8 +5,14 @@ const userscriptHeader* = genMetadataBlock(
   description = "A dirty and free way to use ChatGPT in Nim",
   author = "Thiago Navarro",
   version = "0.1.0",
+  match = [
+    "https://chat.openai.com/*"
+  ],
   runAt = GmRunAt.docIdle,
   grant = [
     GmPermitions.xmlHttpRequest,
   ],
+  require = [
+    "https://unpkg.com/turndown/dist/turndown.js",
+  ]
 )